+++ src/pages/privileges/outlets/users/tabs/invitations/index.tsx
@@ -1,35 +1,20 @@
-import {
-  Table,
-  useMediaQuery,
-  SectionMessage,
-  Stack,
-  Icon,
-} from "@inube/design-system";
-import { Link } from "react-router-dom";
-import { MdOutlineAssignmentTurnedIn } from "react-icons/md";
+import { Table } from "@inube/design-system";
+import { useMediaQuery } from "@inube/design-system";
 import { useEffect, useState } from "react";
+
 import { EMessageType, IMessage } from "@src/types/messages.types";
-import { EAppearance } from "@src/types/colors.types";
-import { deleteItemData, getAll } from "@mocks/utils/dataMock.service";
+import { getAll } from "@mocks/utils/dataMock.service";
 import { LoadingApp } from "@pages/login/outlets/LoadingApp";
 import { IInvitationsEntry } from "@src/services/users/invitation.types";
-import { resendInvitationMessages } from "../../config/resendInvitationUser.config";
+
 import {
   invitationsTableBreakpoints,
   invitationsTableTitles,
 } from "../../config/invitationsTable.config";
-import { DeleteInvitation } from "./DeleteInvitation";
-import { ResendInvitation } from "./ResendInvitation";
-import { StyledMessageContainer } from "./styles";
-import { deleteInvitationModal } from "./DeleteInvitation/config/deleteInvitation.config";
-
-const initialMessageState: IMessage = {
-  show: false,
-  title: "",
-  description: "",
-  icon: <></>,
-  appearance: "" as EAppearance,
-};
+import { resendInvitationMessages } from "../../config/resendInvitationUser.config";
+import { actionsConfigInvitation } from "./config/dataInvitation";
+import { RenderMessage } from "@src/components/feedback/RenderMessage";
+import { IMessageState } from "../../types/forms.types";
 
 interface InvitationsTabProps {
   searchText: string;
@@ -37,11 +22,12 @@ interface InvitationsTabProps {
 
 function InvitationsTab(props: InvitationsTabProps) {
   const { searchText } = props;
-  const [message, setMessage] = useState(initialMessageState);
+  const [message, setMessage] = useState<IMessageState>({
+    visible: false,
+  });
   const [loading, setLoading] = useState(true);
-  const [isHovered, setIsHovered] = useState(false);
   const [invitations, setInvitations] = useState<IInvitationsEntry[]>([]);
-
+  const [isHovered, setIsHovered] = useState(false);
   const smallScreen = useMediaQuery("(max-width: 850px)");
 
   useEffect(() => {
@@ -58,68 +44,6 @@ function InvitationsTab(props: InvitationsTabProps) {
         setLoading(false);
       });
   }, [invitations]);
-
-  const invitationsTableActions = [
-    {
-      id: "1",
-      actionName: "Completar",
-      content: ({
-        invitationId,
-        status,
-      }: {
-        invitationId: string;
-        status: string;
-      }) => (
-        <Link to={`complete-invitation/${invitationId}`}>
-          <Stack
-            onMouseEnter={() => setIsHovered(true)}
-            onMouseLeave={() => setIsHovered(false)}
-          >
-            <Icon
-              appearance={isHovered ? "primary" : "dark"}
-              parentHover={isHovered ? true : false}
-              icon={<MdOutlineAssignmentTurnedIn />}
-              disabled={status === "Sent"}
-              cursorHover
-            />
-          </Stack>
-        </Link>
-      ),
-      type: "gray",
-    },
-
-    {
-      id: "2",
-      actionName: "Reenviar",
-      content: (invitation: IInvitationsEntry) => (
-        <ResendInvitation
-          invitation={invitation}
-          handleResendInvitation={() => handleResendInvitation(invitation)}
-          showComplete={smallScreen}
-        />
-      ),
-      type: "primary",
-    },
-    {
-      id: "3",
-      actionName: "Eliminar",
-      content: ({
-        customerId,
-        userName,
-      }: {
-        customerId: string;
-        userName: string;
-      }) => (
-        <DeleteInvitation
-          deleteInvitation={deleteInvitationModal}
-          linixInvitation={customerId}
-          userNameInvitation={userName}
-          handleDeleteInvitation={deleteItemData}
-        />
-      ),
-      type: "error",
-    },
-  ];
 
   const handleResendInvitation = (invitation: IInvitationsEntry) => {
     let messageType = EMessageType.SUCCESS;
@@ -142,18 +66,16 @@ function InvitationsTab(props: InvitationsTabProps) {
   };
 
   const handleShowMessage = (message: IMessage) => {
-    const { title, description, icon, appearance } = message;
     setMessage({
-      show: true,
-      title,
-      description,
-      icon,
-      appearance,
+      visible: true,
+      data: message,
     });
   };
 
   const handleCloseMessage = () => {
-    setMessage(initialMessageState);
+    setMessage({
+      visible: false,
+    });
   };
 
   return (
@@ -165,25 +87,23 @@ function InvitationsTab(props: InvitationsTabProps) {
           id="portal"
           titles={invitationsTableTitles}
           entries={invitations}
-          actions={invitationsTableActions}
+          actions={actionsConfigInvitation(
+            isHovered,
+            handleResendInvitation,
+            setIsHovered,
+            smallScreen
+          )}
           breakpoints={invitationsTableBreakpoints}
           filter={searchText}
           modalTitle="Invitación"
         />
       )}
-      {message.show && (
-        <StyledMessageContainer>
-          <Stack justifyContent="flex-end" width="100%">
-            <SectionMessage
-              title={message.title}
-              description={message.description}
-              icon={message.icon}
-              appearance={message.appearance}
-              duration={4000}
-              closeSectionMessage={handleCloseMessage}
-            />
-          </Stack>
-        </StyledMessageContainer>
+      {message.visible && (
+        <RenderMessage
+          message={message}
+          handleCloseMessage={handleCloseMessage}
+          onMessageClosed={handleCloseMessage}
+        />
       )}
     </>
   );
