--- src/pages/privileges/outlets/users/tabs/invitations/styles.ts
@@ -1,10 +1,0 @@
-import styled from "styled-components";
-
-const StyledMessageContainer = styled.div`
-  position: fixed;
-  bottom: 18px;
-  right: 75px;
-  z-index: 2;
-`;
-
-export { StyledMessageContainer };
